feat: implement core functionality for Meeting Summarization Gateway; add routers, service status model, logging, and configuration settings
