+++ backend/gateway/models/service_status.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class ServiceStatus(BaseModel):
+    service: str
+    status:  str
+    message: str = ""
