+++ backend/gateway/routers/root.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter 
+
+router = APIRouter()
+
+@router.get("/", response_model=dict)
+def root():
+    return {"status": "gateway running"}
